# src/configs/apidocs/api_docs_v1.yaml
openapi: 3.0.1
info:
  title: Human Resource Management System
  contact:
    email: dothienbinh13@gmail.com
  version: 1.0.0
  servers:
    - url: http://localhost:3000/
  tags:
    - name: users
      description: Everythong about users
    - name: jobs
      description: Everythong about job form
    - name: reviews
      description: Everythong about review form
paths:
...
  /user/{id}/edit:
    get:
      tags:
        - users
      summary: user get infor user
      description: users get their own information to edit
      operationId: getEditUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "ok"
          content:
            application/json:
              example:
                Code: 0
                message: "OK"
                user:
                  {
                    MaNV: B18DCAT020,
                    FirstName: Do,
                    LastName: Binh,
                    Email: dothienbinh@gmail,
                    UserName: BinhDo20,
                    Phone: "123456",
                    Avatar: "http://localhost:3000/avatar-1660042035597.jpg",
                    Gender: "0",
                    CMND: "1234",
                    BHXH: "1234",
                    Address: Hn,
                  }
        400:
          content:
            application/json:
              example:
                errCode: -1
                status: "fail"
                typeError: Http error
                statusError: Bad request
                message: request fail !!!
...
components:
...
  securitySchemes:
    ACCESS_TOKEN: # arbitrary name for the security scheme
      type: http
      scheme: basic
      bearerFormat: JWT
      in: cookie